Add files via upload
ETExplorer is designed for users who find File Explorer complicated in newer versions of Windows. The application, which allows you to easily review and organize your files and folders on your computer, also includes a calendar and different theme options. In short, ETExplorer is a simple but multi-purpose file manager. You can compile it with Visual C++ 2010 and above.